Hard-code path for config file directory
Since the path is used on build time it shouldn't be configurable
runtime. Unfortunately this makes it harder to share the value for the
path between the docker image and the application server but it is
preferable to have it hard-coded in two different places than creating a
false image of configurability.
Also unfortunate is that docker files don't have an instruction to
define constants. ARG would not be configurable runtime but would also
not be reachable by the application server like environment variables
are.